src/requirements.txt
Flask==3.0.3
pydantic==2.8.1
setuptools==73.0.1
# gunicorn==23.0.0
requests==2.32.3
boto3==1.34.127
python-dotenv
streamlit==1.28.2
streamlit_feedback==0.1.3
pyyaml==6.0.1

